# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comcin API",
        "description": "API documentation for the Coalition of Micro-lending and Cooperative Institutions in Nigeria (Comcin)",
        "contact": {
            "email": "support@comcin.ng"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/member/support/tickets": {
            "get": {
                "tags": [
                    "Member - Support Tickets"
                ],
                "summary": "Get All Support Tickets",
                "description": "Retrieves all support tickets for the authenticated member.",
                "operationId": "4d93c68e2ac2867710b830a6d608e6c4",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of support tickets."
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Member - Support Tickets"
                ],
                "summary": "Create Support Ticket",
                "description": "Creates a new support ticket for the authenticated member.",
                "operationId": "22014d96eeb61386f642312f8f8f0c51",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "subject",
                                    "message"
                                ],
                                "properties": {
                                    "subject": {
                                        "type": "string",
                                        "example": "Issue with Membership"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "I am having issues with my membership renewal."
                                    },
                                    "attachment": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Support ticket created successfully."
                    },
                    "400": {
                        "description": "Bad Request - Validation errors or missing fields."
                    },
                    "401": {
                        "description": "Unauthorized - User not authenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/member/support/tickets/{uuid}": {
            "get": {
                "tags": [
                    "Member - Support Tickets"
                ],
                "summary": "Get Support Ticket",
                "description": "Retrieves a specific support ticket by UUID for the authenticated member.",
                "operationId": "4d7a6a870cc3c51be00ab26d75b8f3e6",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the support ticket.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of the support ticket."
                    },
                    "404": {
                        "description": "Support ticket not found."
                    },
                    "401": {
                        "description": "Unauthorized - User not authenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format **Bearer &lt;token&gt;**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}